@@ -6,12 +6,22 @@ export interface ROOTS {
     idx: number
 }
 
-export const KnownTransaction = 'knownTransaction'
-export const UnknownTransaction = 'unknownTransaction'
-export const EmptyBlock = 'emptyBlock'
-export const NewTransaction = 'newTransaction'
-export const NewBlock = 'newBlock'
-export const NewCall = 'newCall'
+export const KNOWN_TRANSACTION = 'knownTransaction'
+export const UNKNOWN_TRANSACTION = 'unknownTransaction'
+export const EMPTY_BLOCK = 'emptyBlock'
+export const NEW_TRANSACTION = 'newTransaction'
+export const NEW_BLOCK = 'newBlock'
+export const NEW_CALL = 'newCall'
+
+export const HTML = 'html'
+export const LOG = 'log'
+export const INFO = 'info'
+export const WARN = 'warn'
+export const ERROR = 'error'
+export const SCRIPT = 'script'
+export const CLEAR_CONSOLE = 'clearconsole'
+export const LISTEN_ON_NETWORK = 'listenOnNetWork'
+export const CMD_HISTORY = 'cmdHistory'
 
 export interface RemixUiTerminalProps {
     plugin: any
